backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["jackson.wen <jackson.wen@xm.wonder.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.112.2"}
sqlalchemy = {extras = ["postgresql-asyncpg"], version = "^2.0.32"}
alembic = "^1.13.2"
redis = "^5.0.8"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlmodel = "^0.0.21"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
